requirements.txt
flask==3.0.2
google-generativeai==0.3.2
werkzeug==3.0.1
kloppy==3.0.0 